requirements.txt
amqp==1.4.7
redis==2.10.5
Flask==0.10.1
gunicorn==19.4.5
gevent==1.0.2
celery==3.1.23
biopython==1.65
numpy==1.14.0
nose==1.3.7
coverage==4.2
kombu==3.0.37
watchdog==0.8.3
wsgiref==0.1.2
filelock==0.2.2
zeep==2.0.0
